:recycle: restructure user store
