+++ src/features/user/constants.ts
@@ -1,0 +1,5 @@
+export const USER_STATUS = {
+  notLoggedIn: 0,
+  guest: 1,
+  member: 2,
+};
